+++ Wireframe-Admob-All-IN-ONE/app/src/main/java/com/admob/android/ads/withwireframe/bestpractices/interfaces/OnShowAdCompleteListener.kt
@@ -1,0 +1,5 @@
+package com.admob.android.ads.withwireframe.bestpractices.interfaces
+
+interface OnShowAdCompleteListener {
+    fun onShowAdComplete()
+}