+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## `12 Jun 2025`
+
+- Wireframe App Open Ad implemented
+
 ## `05 Jun 2025`
 
 - Added AdMob App Open Ad